rename audio handler to SoundDeck, add >true
SoundDeck also comes with some more improvements. (i.e, it works now. and pools objects! very awesome)
>true is REALLY slow, but that's okay. caretaker archive is back up! i should probably rename that bot.
replace longs in economy with decimals "0.00m", so that jells act more like dollars than cents/yen
"fix" connect 4 solving, just had to revert :cry:
save all logs to a file, where all log files over 1 day old get deleted

+++ scripts/Economy/EconomyHandler.cs
@@ -4,9 +4,9 @@ namespace CaretakerNET.Economy
 {
     public class EconomyHandler
     {
-        public class Item(long price, string name, string desc)
+        public class Item(decimal price, string name, string desc)
         {
-            public long Price = price;
+            public decimal Price = price;
             public string Name = name;
             public string Desc = desc;
         }
